+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/rule/xpath/SaxonXPathRuleQuery.java
@@ -14,12 +14,9 @@ import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.ast.xpath.saxon.DocumentNode;
 import net.sourceforge.pmd.lang.ast.xpath.saxon.ElementNode;
 import net.sourceforge.pmd.lang.xpath.Initializer;
-import net.sourceforge.pmd.properties.BooleanProperty;
-import net.sourceforge.pmd.properties.EnumeratedProperty;
-import net.sourceforge.pmd.properties.IntegerProperty;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
-import net.sourceforge.pmd.properties.StringProperty;
 
+import net.sf.saxon.om.Item;
 import net.sf.saxon.om.ValueRepresentation;
 import net.sf.saxon.sxpath.AbstractStaticContext;
 import net.sf.saxon.sxpath.IndependentContext;
@@ -29,27 +26,36 @@ import net.sf.saxon.sxpath.XPathExpression;
 import net.sf.saxon.sxpath.XPathStaticContext;
 import net.sf.saxon.sxpath.XPathVariable;
 import net.sf.saxon.trans.XPathException;
+import net.sf.saxon.value.AtomicValue;
+import net.sf.saxon.value.BigIntegerValue;
 import net.sf.saxon.value.BooleanValue;
+import net.sf.saxon.value.DoubleValue;
+import net.sf.saxon.value.EmptySequence;
+import net.sf.saxon.value.FloatValue;
 import net.sf.saxon.value.Int64Value;
+import net.sf.saxon.value.SequenceExtent;
 import net.sf.saxon.value.StringValue;
+import net.sf.saxon.value.UntypedAtomicValue;
+
 
 /**
  * This is a Saxon based XPathRule query.
  */
 public class SaxonXPathRuleQuery extends AbstractXPathRuleQuery {
 
-    // Mapping from Node name to applicable XPath queries
-    private XPathExpression xpathExpression;
-    private List<XPathVariable> xpathVariables;
-
     private static final int MAX_CACHE_SIZE = 20;
     private static final Map<Node, DocumentNode> CACHE = new LinkedHashMap<Node, DocumentNode>(MAX_CACHE_SIZE) {
         private static final long serialVersionUID = -7653916493967142443L;
+
 
         protected boolean removeEldestEntry(final Map.Entry<Node, DocumentNode> eldest) {
             return size() > MAX_CACHE_SIZE;
         }
     };
+    // Mapping from Node name to applicable XPath queries
+    private XPathExpression xpathExpression;
+    private List<XPathVariable> xpathVariables;
+
 
     /**
      * {@inheritDoc}
@@ -58,6 +64,7 @@ public class SaxonXPathRuleQuery extends AbstractXPathRuleQuery {
     public boolean isSupportedVersion(String version) {
         return XPATH_1_0_COMPATIBILITY.equals(version) || XPATH_2_0.equals(version);
     }
+
 
     /**
      * {@inheritDoc}
@@ -83,31 +90,7 @@ public class SaxonXPathRuleQuery extends AbstractXPathRuleQuery {
                 String name = xpathVariable.getVariableQName().getLocalName();
                 for (Map.Entry<PropertyDescriptor<?>, Object> entry : super.properties.entrySet()) {
                     if (name.equals(entry.getKey().name())) {
-                        PropertyDescriptor<?> propertyDescriptor = entry.getKey();
-                        Object value = entry.getValue();
-                        ValueRepresentation valueRepresentation;
-
-                        // TODO Need to handle null values?
-                        // TODO Need to handle more PropertyDescriptors, is
-                        // there an easy factory in Saxon we can use for this?
-                        if (propertyDescriptor instanceof StringProperty) {
-                            valueRepresentation = new StringValue((String) value);
-                        } else if (propertyDescriptor instanceof BooleanProperty) {
-                            valueRepresentation = BooleanValue.get((Boolean) value);
-                        } else if (propertyDescriptor instanceof IntegerProperty) {
-                            valueRepresentation = Int64Value.makeIntegerValue((Integer) value);
-                        } else if (propertyDescriptor instanceof EnumeratedProperty) {
-                            if (value instanceof String) {
-                                valueRepresentation = new StringValue((String) value);
-                            } else {
-                                throw new RuntimeException(
-                                        "Unable to create ValueRepresentaton for non-String EnumeratedProperty value: "
-                                                + value);
-                            }
-                        } else {
-                            throw new RuntimeException("Unable to create ValueRepresentaton for PropertyDescriptor: "
-                                    + propertyDescriptor);
-                        }
+                        ValueRepresentation valueRepresentation = getRepresentation(entry.getKey(), entry.getValue());
                         xpathDynamicContext.setVariable(xpathVariable, valueRepresentation);
                     }
                 }
@@ -122,6 +105,24 @@ public class SaxonXPathRuleQuery extends AbstractXPathRuleQuery {
         }
         return results;
     }
+
+
+    private ValueRepresentation getRepresentation(PropertyDescriptor<?> descriptor, Object value) {
+        if (descriptor.isMultiValue()) {
+            List<?> val = (List<?>) value;
+            if (val.isEmpty()) {
+                return EmptySequence.getInstance();
+            }
+            Item[] converted = new Item[val.size()];
+            for (int i = 0; i < val.size(); i++) {
+                converted[i] = getAtomicRepresentation(val.get(i));
+            }
+            return new SequenceExtent(converted);
+        } else {
+            return getAtomicRepresentation(value);
+        }
+    }
+    
 
     private DocumentNode getDocumentNode(Node node) {
         // Get the root AST node
@@ -141,6 +142,7 @@ public class SaxonXPathRuleQuery extends AbstractXPathRuleQuery {
         }
         return documentNode;
     }
+
 
     private void initializeXPathExpression() {
         if (xpathExpression != null) {
@@ -176,6 +178,38 @@ public class SaxonXPathRuleQuery extends AbstractXPathRuleQuery {
             xpathExpression = xpathEvaluator.createExpression(super.xpath);
         } catch (XPathException e) {
             throw new RuntimeException(e);
+        }
+    }
+
+
+    /**
+     * Gets the Saxon representation of the parameter, if its type corresponds 
+     * to an XPath 2.0 atomic datatype.
+     *
+     * @param value The value to convert
+     *
+     * @return The converted AtomicValue
+     */
+    public static AtomicValue getAtomicRepresentation(Object value) {
+        if (value == null) {
+            return UntypedAtomicValue.ZERO_LENGTH_UNTYPED;
+        } else if (value instanceof String) {
+            return new StringValue((String) value);
+        } else if (value instanceof Boolean) {
+            return BooleanValue.get((Boolean) value);
+        } else if (value instanceof Integer) {
+            return Int64Value.makeIntegerValue((Integer) value);
+        } else if (value instanceof Long) {
+            return new BigIntegerValue((Long) value);
+        } else if (value instanceof Double) {
+            return new DoubleValue((Double) value);
+        } else if (value instanceof Character) {
+            return new StringValue(value.toString());
+        } else if (value instanceof Float) {
+            return new FloatValue((Float) value);
+        } else {
+            // We could maybe use UntypedAtomicValue
+            throw new RuntimeException("Unable to create ValueRepresentation for value of type: " + value.getClass());
         }
     }
 }